fix: useEffect in app.js, change let to const in login.js/register.js, base state in addplacepopup.js
